@@ -1,40 +1,39 @@
-import abc
 import logging
+from abc import ABC, abstractmethod
 from typing import Any
 
 import httpx
 import xmltodict
-from addict import Dict
 
 from hikcamerabot.clients.hikvision import HikvisionAPIClient
 from hikcamerabot.clients.hikvision.enums import EndpointAddr
+from hikcamerabot.constants import XML_HEADERS
 from hikcamerabot.exceptions import HikvisionAPIError
 
 
-class AbstractEndpoint(metaclass=abc.ABCMeta):
+class AbstractEndpoint(ABC):
     """API Endpoint class.
 
     Used to decompose API methods since they are too complex to store in one API class.
     """
 
     _XML_PAYLOAD_TPL: str | None = None
-    _XML_HEADERS = {'Content-Type': 'application/xml'}
 
     def __init__(self, api_client: HikvisionAPIClient) -> None:
         self._api_client = api_client
         self._log = logging.getLogger(self.__class__.__name__)
 
-    @abc.abstractmethod
+    @abstractmethod
     async def __call__(self, *args, **kwargs) -> Any:
         """Real API call starts here."""
 
-    async def _get_channel_capabilities(self) -> Dict:
+    async def _get_channel_capabilities(self) -> dict[str, Any]:
         response = await self._api_client.request(
             method='GET',
             endpoint=EndpointAddr.CHANNEL_CAPABILITIES,
-            headers=self._XML_HEADERS,
+            headers=XML_HEADERS,
         )
-        return Dict(xmltodict.parse(response.text))
+        return xmltodict.parse(response.text)
 
     def _validate_xml_response(self, response: httpx.Response) -> None:
         xml_text = response.text
@@ -45,7 +44,7 @@ class AbstractEndpoint(metaclass=abc.ABCMeta):
             err_msg = f'Failed to parse response XML: {err}'
             self._log.error(err)
             self._log.debug(xml_text)
-            raise HikvisionAPIError(err_msg)
+            raise HikvisionAPIError(err_msg) from err
         if is_err:
             err_msg = 'Camera returned failed errored XML'
             self._log.error(err_msg)
