@@ -1,9 +1,9 @@
 from collections.abc import AsyncGenerator
 from io import BytesIO
+from typing import Any
 from urllib.parse import urljoin
 
 import httpx
-from addict import Dict
 
 from hikcamerabot.clients.hikvision.endpoints.abstract import AbstractEndpoint
 from hikcamerabot.clients.hikvision.endpoints.config_switch import CameraConfigSwitch
@@ -14,12 +14,13 @@ from hikcamerabot.clients.hikvision.enums import (
     OverexposeSuppressEnabledType,
     OverexposeSuppressType,
 )
-from hikcamerabot.constants import CONN_TIMEOUT, Detection
+from hikcamerabot.constants import CONN_TIMEOUT, XML_HEADERS
+from hikcamerabot.enums import DetectionType
 from hikcamerabot.exceptions import APIRequestError
 
 
 class IrcutFilterEndpoint(AbstractEndpoint):
-    _XML_PAYLOAD_TPL = (
+    _XML_PAYLOAD_TPL: str = (
         '<IrcutFilter>'
         '<IrcutFilterType>{filter_type}</IrcutFilterType>'
         '<nightToDayFilterLevel>{night_to_day_filter_level}</nightToDayFilterLevel>'
@@ -32,8 +33,10 @@ class IrcutFilterEndpoint(AbstractEndpoint):
         try:
             response = await self._api_client.request(
                 endpoint=EndpointAddr.IRCUT_FILTER,
-                headers=self._XML_HEADERS,
-                data=self._build_payload(filter_type, current_capabilities),
+                headers=XML_HEADERS,
+                data=self._build_payload(
+                    filter_type=filter_type, current_capabilities=current_capabilities
+                ),
                 method='PUT',
             )
         except APIRequestError:
@@ -44,14 +47,13 @@ class IrcutFilterEndpoint(AbstractEndpoint):
         self._validate_xml_response(response)
 
     def _build_payload(
-        self, filter_type: IrcutFilterType, current_capabilities: Dict
+        self, filter_type: IrcutFilterType, current_capabilities: dict[str, Any]
     ) -> str:
-        level: str = (
-            current_capabilities.ImageChannel.IrcutFilter.nightToDayFilterLevel['#text']
-        )
-        time: str = current_capabilities.ImageChannel.IrcutFilter.nightToDayFilterTime[
-            '#text'
+        ircut_filter: dict[str, Any] = current_capabilities['ImageChannel'][
+            'IrcutFilter'
         ]
+        level: str = ircut_filter['nightToDayFilterLevel']['#text']
+        time: str = ircut_filter['nightToDayFilterTime']['#text']
         return self._XML_PAYLOAD_TPL.format(
             filter_type=filter_type.value,
             night_to_day_filter_level=level,
@@ -60,7 +62,7 @@ class IrcutFilterEndpoint(AbstractEndpoint):
 
 
 class ExposureEndpoint(AbstractEndpoint):
-    _XML_PAYLOAD_TPL = (
+    _XML_PAYLOAD_TPL: str = (
         '<Exposure>'
         '<ExposureType>{exposure_type}</ExposureType>'
         '<OverexposeSuppress>'
@@ -73,10 +75,10 @@ class ExposureEndpoint(AbstractEndpoint):
 
     async def __call__(
         self,
-        exposure_type: ExposureType = None,
-        overexpose_suppress_enabled: OverexposeSuppressEnabledType = None,
-        overexposure_suppress_type: OverexposeSuppressType = None,
-        distance_level: int = None,
+        exposure_type: ExposureType | None = None,
+        overexpose_suppress_enabled: OverexposeSuppressEnabledType | None = None,
+        overexposure_suppress_type: OverexposeSuppressType | None = None,
+        distance_level: int | None = None,
     ) -> None:
         kwargs = {
             'exposure_type': exposure_type,
@@ -85,16 +87,18 @@ class ExposureEndpoint(AbstractEndpoint):
             'distance_level': distance_level,
         }
         kwargs_len = len(kwargs)
-        filtered_kwargs = Dict({k: v for k, v in kwargs if v is not None})
+        filtered_kwargs = {k: v for k, v in kwargs if v is not None}
 
-        current_capabilities = None
+        current_capabilities: dict[str, Any] | None = None
         if len(filtered_kwargs) != kwargs_len:
             current_capabilities = await self._get_channel_capabilities()
         try:
             response = await self._api_client.request(
                 endpoint=EndpointAddr.IRCUT_FILTER,
-                headers=self._XML_HEADERS,
-                data=self._build_payload(filtered_kwargs, current_capabilities),
+                headers=XML_HEADERS,
+                data=self._build_payload(
+                    kwargs=filtered_kwargs, current_capabilities=current_capabilities
+                ),
                 method='PUT',
             )
         except APIRequestError:
@@ -102,23 +106,25 @@ class ExposureEndpoint(AbstractEndpoint):
             raise
         self._validate_xml_response(response)
 
-    def _build_payload(self, kwargs: Dict, current_capabilities: Dict | None) -> str:
+    def _build_payload(
+        self, kwargs: dict, current_capabilities: dict[str, Any] | None
+    ) -> str:
         return self._XML_PAYLOAD_TPL.format(
             exposure_type=kwargs.get(
                 'exposure_type',
-                current_capabilities.exposure_type,
+                current_capabilities['exposure_type'],
             ),
             overexpose_suppress_enabled=kwargs.get(
                 'overexpose_suppress_enabled',
-                current_capabilities.overexpose_suppress_enabled,
+                current_capabilities['overexpose_suppress_enabled'],
             ),
             overexposure_suppress_type=kwargs.get(
                 'overexposure_suppress_type',
-                current_capabilities.overexposure_suppress_type,
+                current_capabilities['overexposure_suppress_type'],
             ),
             distance_level=kwargs.get(
                 'distance_level',
-                current_capabilities.distance_level,
+                current_capabilities['distance_level'],
             ),
         )
 
@@ -136,17 +142,17 @@ class TakeSnapshotEndpoint(AbstractEndpoint):
 
 
 class AlertStreamEndpoint(AbstractEndpoint):
-    _METHOD = 'GET'
+    _METHOD: str = 'GET'
 
-    async def __call__(self) -> AsyncGenerator[str, None]:
+    async def __call__(self) -> AsyncGenerator[str]:
         # TODO: Rewrite this.
         url = urljoin(
             f'{self._api_client.host}:{self._api_client.port}',
-            EndpointAddr.ALERT_STREAM.value,
+            EndpointAddr.ALERT_STREAM,
         )
         timeout = httpx.Timeout(CONN_TIMEOUT, read=300)
         response: httpx.Response
-        self._log.debug('Alert Stream Request: %s - %s', self._METHOD, url)
+        self._log.debug('Alert StreamType Request: %s - %s', self._METHOD, url)
         async with self._api_client.session.stream(
             self._METHOD, url, timeout=timeout
         ) as response:
@@ -160,7 +166,7 @@ class SwitchEndpoint(AbstractEndpoint):
         super().__init__(*args, **kwargs)
         self._switch = CameraConfigSwitch(api_client=self._api_client)
 
-    async def __call__(self, trigger: Detection, state: bool) -> str | None:
+    async def __call__(self, trigger: DetectionType, state: bool) -> str | None:
         """Switch method to enable/disable Hikvision functions.
 
         :param state: Boolean value indicating on/off switch state.
