@@ -1,4 +1,4 @@
-import os
+from pathlib import Path
 
 from hikcamerabot.common.video.tasks.abstract import AbstractFfBinaryTask
 from hikcamerabot.utils.process import get_stdout_stderr
@@ -7,7 +7,7 @@ from hikcamerabot.utils.process import get_stdout_stderr
 class MakeThumbnailTask(AbstractFfBinaryTask):
     _CMD = 'ffmpeg -y -loglevel error -i {filepath} -vframes 1 -q:v 31 {thumbpath}'
 
-    def __init__(self, thumbnail_path: str, *args, **kwargs) -> None:
+    def __init__(self, thumbnail_path: Path, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
         self._thumbnail_path = thumbnail_path
 
@@ -39,11 +39,11 @@ class MakeThumbnailTask(AbstractFfBinaryTask):
 
         For example, zero-size thumbnail could be created when no space left on device.
         """
-        if not os.path.exists(self._thumbnail_path):
+        if self._thumbnail_path.exists():
             return
 
         self._log.info('Cleaning up errored "%s"', self._thumbnail_path)
         try:
-            os.remove(self._thumbnail_path)
+            self._thumbnail_path.unlink()
         except Exception:
             self._log.exception('Cleanup failed for errored "%s"', self._thumbnail_path)
