@@ -1,98 +1,71 @@
-from marshmallow import INCLUDE, Schema, validates_schema
-from marshmallow import fields as f
-
-from hikcamerabot.config.schemas.validators import (
-    int_min_0,
-    int_min_1,
-    int_min_minus_1,
-    non_empty_str,
-)
+from hikcamerabot.config.schemas._types import IntMin0, IntMin1, IntMinus1
+from hikcamerabot.config.schemas.abstract import StrictBaseModel
+from hikcamerabot.enums import StreamType
 
 
-class BaseTemplate(Schema):
-    _inner_validation_schema_cls: type[Schema] | None = None
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self._inner_validation_schema = self._inner_validation_schema_cls()
-        self._template_validator = f.String(required=True, validate=non_empty_str)
-
-    @validates_schema
-    def validate_all(self, data: dict, **kwargs) -> None:
-        for tpl_name, tpl_conf in data.items():
-            self._template_validator.validate(tpl_name)
-            self._inner_validation_schema.load(tpl_conf)
-
-    class Meta:
-        unknown = INCLUDE
+class YoutubeSchema(StrictBaseModel):
+    channel: IntMin1
+    restart_period: IntMin1
+    restart_pause: IntMin0
+    url: str
+    key: str
 
 
-class Youtube(BaseTemplate):
-    class _Youtube(Schema):
-        channel = f.Integer(required=True, validate=int_min_1)
-        restart_period = f.Integer(required=True, validate=int_min_1)
-        restart_pause = f.Integer(required=True, validate=int_min_0)
-        url = f.String(required=True, validate=non_empty_str)
-        key = f.String(required=True, validate=non_empty_str)
-
-    _inner_validation_schema_cls = _Youtube
+class TelegramSchema(StrictBaseModel):
+    channel: IntMin1
+    restart_period: IntMin1
+    restart_pause: IntMin0
+    url: str
+    key: str
 
 
-class Telegram(BaseTemplate):
-    class _Telegram(Schema):
-        channel = f.Integer(required=True, validate=int_min_1)
-        restart_period = f.Integer(required=True, validate=int_min_1)
-        restart_pause = f.Integer(required=True, validate=int_min_0)
-        url = f.String(required=True, validate=non_empty_str)
-        key = f.String(required=True, validate=non_empty_str)
-
-    _inner_validation_schema_cls = _Telegram
+class SrsSchema(StrictBaseModel):
+    channel: IntMin1
+    sub_channel: IntMin1
+    restart_period: IntMinus1
+    restart_pause: IntMin0
+    url: str
 
 
-class Srs(BaseTemplate):
-    class _Srs(Schema):
-        channel = f.Integer(required=True, validate=int_min_1)
-        sub_channel = f.Integer(required=True, validate=int_min_1)
-        restart_period = f.Integer(required=True, validate=int_min_minus_1)
-        restart_pause = f.Integer(required=True, validate=int_min_0)
-        url = f.String(required=True, validate=non_empty_str)
-
-    _inner_validation_schema_cls = _Srs
+class DvrSchema(StrictBaseModel):
+    channel: IntMin1
+    sub_channel: IntMin1
+    restart_period: IntMinus1
+    restart_pause: IntMin0
+    segment_time: IntMin1
 
 
-class Dvr(BaseTemplate):
-    class _Dvr(Schema):
-        channel = f.Integer(required=True, validate=int_min_1)
-        sub_channel = f.Integer(required=True, validate=int_min_1)
-        restart_period = f.Integer(required=True, validate=int_min_minus_1)
-        restart_pause = f.Integer(required=True, validate=int_min_0)
-        segment_time = f.Integer(required=True, validate=int_min_1)
-
-    _inner_validation_schema_cls = _Dvr
-
-
-class Icecast(BaseTemplate):
-    class _Icecast(Schema):
-        class _IceStream(Schema):
-            ice_genre = f.String(required=True, validate=non_empty_str)
-            ice_name = f.String(required=True, validate=non_empty_str)
-            ice_description = f.String(required=True, validate=non_empty_str)
-            ice_public = f.Integer(required=True)
-            url = f.String(required=True, validate=non_empty_str)
-            password = f.String(required=True, validate=non_empty_str)
-            content_type = f.String(required=True, validate=non_empty_str)
-
-        channel = f.Integer(required=True, validate=int_min_1)
-        restart_period = f.Integer(required=True, validate=int_min_1)
-        restart_pause = f.Integer(required=True, validate=int_min_0)
-        ice_stream = f.Nested(_IceStream(), required=True)
-
-    _inner_validation_schema_cls = _Icecast
+class IceStreamSchema(StrictBaseModel):
+    ice_genre: str
+    ice_name: str
+    ice_description: str
+    ice_public: int
+    url: str
+    password: str
+    content_type: str
 
 
-class Livestream(Schema):
-    youtube = f.Nested(Youtube(), required=True)
-    telegram = f.Nested(Telegram(), required=True)
-    icecast = f.Nested(Icecast(), required=True)
-    srs = f.Nested(Srs(), required=True)
-    dvr = f.Nested(Dvr(), required=True)
+class IcecastSchema(StrictBaseModel):
+    channel: IntMin1
+    restart_period: IntMin1
+    restart_pause: IntMin0
+    ice_stream: IceStreamSchema
+
+
+type LiveStreamTplType = dict[
+    str, YoutubeSchema | TelegramSchema | IcecastSchema | SrsSchema | DvrSchema
+]
+
+
+class LivestreamTemplatesSchema(StrictBaseModel):
+    youtube: dict[str, YoutubeSchema]
+    telegram: dict[str, TelegramSchema]
+    icecast: dict[str, IcecastSchema]
+    srs: dict[str, SrsSchema]
+    dvr: dict[str, DvrSchema]
+
+    def get_tpl_by_name(self, name: StreamType) -> LiveStreamTplType:
+        try:
+            return getattr(self, name.value.lower())
+        except AttributeError:
+            raise ValueError(f'No template with name {name}') from None
