@@ -6,12 +6,13 @@ from functools import wraps
 from typing import TYPE_CHECKING
 
 from emoji import emojize
-from pyrogram.types import Message
 
 from hikcamerabot.constants import CMD_CAM_ID_REGEX
 from hikcamerabot.utils.shared import get_user_info
 
 if TYPE_CHECKING:
+    from pyrogram.types import Message
+
     from hikcamerabot.camerabot import CameraBot
 
 log = logging.getLogger(__name__)
