@@ -1,9 +1,12 @@
-"""Event dispatcher module."""
+"""EventType dispatcher module."""
 
-from hikcamerabot.enums import Event
+from typing import ClassVar
+
+from hikcamerabot.enums import EventType
 from hikcamerabot.event_engine.dispatchers.abstract import AbstractDispatcher
 from hikcamerabot.event_engine.events.abstract import BaseInboundEvent
 from hikcamerabot.event_engine.handlers.inbound import (
+    AbstractTaskEvent,
     TaskAlarmConf,
     TaskDetectionConf,
     TaskIrcutFilterConf,
@@ -14,15 +17,15 @@ from hikcamerabot.event_engine.handlers.inbound import (
 
 
 class InboundEventDispatcher(AbstractDispatcher):
-    """Inbound Event Dispatcher Class."""
+    """Inbound EventType Dispatcher Class."""
 
-    DISPATCH = {
-        Event.CONFIGURE_ALARM: TaskAlarmConf,
-        Event.CONFIGURE_DETECTION: TaskDetectionConf,
-        Event.CONFIGURE_IRCUT_FILTER: TaskIrcutFilterConf,
-        Event.STREAM: TaskStreamConf,
-        Event.TAKE_SNAPSHOT: TaskTakeSnapshot,
-        Event.RECORD_VIDEOGIF: TaskRecordVideoGif,
+    DISPATCH: ClassVar[dict[EventType, AbstractTaskEvent]] = {
+        EventType.CONFIGURE_ALARM: TaskAlarmConf,
+        EventType.CONFIGURE_DETECTION: TaskDetectionConf,
+        EventType.CONFIGURE_IRCUT_FILTER: TaskIrcutFilterConf,
+        EventType.STREAM: TaskStreamConf,
+        EventType.TAKE_SNAPSHOT: TaskTakeSnapshot,
+        EventType.RECORD_VIDEOGIF: TaskRecordVideoGif,
     }
 
     async def dispatch(self, event: BaseInboundEvent) -> None:
