@@ -1,9 +1,12 @@
 """Result dispatcher module."""
 
-from hikcamerabot.enums import Event
+from typing import ClassVar
+
+from hikcamerabot.enums import EventType
 from hikcamerabot.event_engine.dispatchers.abstract import AbstractDispatcher
 from hikcamerabot.event_engine.events.abstract import BaseOutboundEvent
 from hikcamerabot.event_engine.handlers.outbound import (
+    AbstractResultEventHandler,
     ResultAlarmConfHandler,
     ResultAlertSnapshotHandler,
     ResultAlertVideoHandler,
@@ -18,15 +21,15 @@ from hikcamerabot.event_engine.handlers.outbound import (
 class OutboundEventDispatcher(AbstractDispatcher):
     """Outbound (Result) Dispatcher Class."""
 
-    DISPATCH = {
-        Event.ALERT_SNAPSHOT: ResultAlertSnapshotHandler,
-        Event.ALERT_VIDEO: ResultAlertVideoHandler,
-        Event.CONFIGURE_ALARM: ResultAlarmConfHandler,
-        Event.CONFIGURE_DETECTION: ResultDetectionConfHandler,
-        Event.SEND_TEXT: ResultSendTextHandler,
-        Event.STREAM: ResultStreamConfHandler,
-        Event.TAKE_SNAPSHOT: ResultTakeSnapshotHandler,
-        Event.RECORD_VIDEOGIF: ResultRecordVideoGifHandler,
+    DISPATCH: ClassVar[dict[EventType, AbstractResultEventHandler]] = {
+        EventType.ALERT_SNAPSHOT: ResultAlertSnapshotHandler,
+        EventType.ALERT_VIDEO: ResultAlertVideoHandler,
+        EventType.CONFIGURE_ALARM: ResultAlarmConfHandler,
+        EventType.CONFIGURE_DETECTION: ResultDetectionConfHandler,
+        EventType.SEND_TEXT: ResultSendTextHandler,
+        EventType.STREAM: ResultStreamConfHandler,
+        EventType.TAKE_SNAPSHOT: ResultTakeSnapshotHandler,
+        EventType.RECORD_VIDEOGIF: ResultRecordVideoGifHandler,
     }
 
     async def dispatch(self, event: BaseOutboundEvent) -> None:
