@@ -2,13 +2,12 @@
 
 import asyncio
 from collections.abc import AsyncGenerator
-from typing import TYPE_CHECKING
-
-from addict import Dict
+from typing import TYPE_CHECKING, Literal
 
 from hikcamerabot.clients.hikvision import HikvisionAPI
+from hikcamerabot.config.schemas.main_config import AlertSchema
 from hikcamerabot.constants import DETECTION_SWITCH_MAP
-from hikcamerabot.enums import Alarm, Detection, ServiceType
+from hikcamerabot.enums import AlarmType, DetectionType, ServiceType
 from hikcamerabot.exceptions import HikvisionAPIError, ServiceRuntimeError
 from hikcamerabot.services.abstract import AbstractService
 from hikcamerabot.services.alarm.camera.tasks.alarm_monitoring_task import (
@@ -24,22 +23,26 @@ if TYPE_CHECKING:
 class AlarmService(AbstractService):
     """Alarm Service Class."""
 
-    ALARM_TRIGGERS = Detection.choices()
+    ALARM_TRIGGERS = DetectionType.choices()
 
-    type = ServiceType.ALARM
-    name = Alarm.ALARM
+    TYPE: Literal[ServiceType.ALARM] = ServiceType.ALARM
+    NAME: Literal[AlarmType.ALARM] = AlarmType.ALARM
 
     def __init__(
-        self, conf: Dict, api: HikvisionAPI, cam: 'HikvisionCam', bot: 'CameraBot'
-    ):
+        self,
+        conf: AlertSchema,
+        api: HikvisionAPI,
+        cam: 'HikvisionCam',
+        bot: 'CameraBot',
+    ) -> None:
         super().__init__(cam)
         self._conf = conf
         self._api = api
         self.bot = bot
         self.alert_delay: int = conf.delay
-        self._alert_count = 0
+        self._alert_count: int = 0
 
-        self._started = asyncio.Event()
+        self._started: asyncio.Event = asyncio.Event()
 
     @property
     def alert_count(self) -> int:
@@ -56,10 +59,10 @@ class AlarmService(AbstractService):
     @property
     def enabled_in_conf(self) -> bool:
         """Check if any alarm trigger is enabled in conf."""
-        for trigger in self.ALARM_TRIGGERS:
-            if self._conf[trigger].enabled:
-                return True
-        return False
+        return any(
+            self._conf.get_detection_schema_by_type(type_=trigger).enabled
+            for trigger in self.ALARM_TRIGGERS
+        )
 
     async def start(self) -> None:
         """Enable alarm service and enable triggers on physical camera."""
@@ -87,8 +90,8 @@ class AlarmService(AbstractService):
 
     async def _enable_triggers_on_camera(self) -> None:
         for trigger in self.ALARM_TRIGGERS:
-            if self._conf[trigger].enabled:
-                await self.trigger_switch(trigger=Detection(trigger), state=True)
+            if self._conf.get_detection_schema_by_type(type_=trigger).enabled:
+                await self.trigger_switch(trigger=DetectionType(trigger), state=True)
 
     async def stop(self) -> None:
         """Disable alarm."""
@@ -96,18 +99,18 @@ class AlarmService(AbstractService):
             raise ServiceRuntimeError('Alarm alert mode already stopped')
         self._started.clear()
 
-    async def alert_stream(self) -> AsyncGenerator[str, None]:
+    async def alert_stream(self) -> AsyncGenerator[str]:
         """Get Alarm stream from Hikvision Camera."""
         async for chunk in self._api.alert_stream():
             yield chunk
 
-    async def trigger_switch(self, trigger: Detection, state: bool) -> str | None:
+    async def trigger_switch(self, trigger: DetectionType, state: bool) -> str | None:
         """Trigger switch."""
-        full_name: str = DETECTION_SWITCH_MAP[trigger]['name'].value
+        full_name = DETECTION_SWITCH_MAP[trigger]['name']
         self._log.debug('%s %s', 'Enabling' if state else 'Disabling', full_name)
         try:
             return await self._api.switch(trigger=trigger, state=state)
         except HikvisionAPIError as err:
             err_msg = f'{full_name} Switch encountered an error: {err}'
             self._log.error(err_msg)
-            raise ServiceRuntimeError(err_msg)
+            raise ServiceRuntimeError(err_msg) from err
