@@ -5,7 +5,7 @@ from httpx import ConnectError
 from tenacity import retry, retry_if_exception_type, wait_fixed
 
 from hikcamerabot.camera import HikvisionCam
-from hikcamerabot.enums import Detection
+from hikcamerabot.enums import DetectionType
 from hikcamerabot.exceptions import AlarmEventChunkDetectorError, ChunkLoopError
 from hikcamerabot.services.alarm.camera.chunk import (
     AlarmEventChunkDetector,
@@ -17,7 +17,7 @@ from hikcamerabot.services.alarm.camera.notifier import AlarmNotifier
 class NvrAlarmMonitoringTask:
     """NVR Alarm Pusher Class."""
 
-    RETRY_WAIT = 0.5
+    RETRY_WAIT: int = 0.5
 
     def __init__(
         self,
@@ -39,7 +39,7 @@ class NvrAlarmMonitoringTask:
         self._cam_delays = _cam_delays
 
         # Each camera connects to the same NVR with the same credentials.
-        # We need Hikvision API instance just to get the NVR Alert Stream.
+        # We need Hikvision API instance just to get the NVR Alert StreamType.
         self._api = cameras[0]._api  # noqa: SLF001
 
         self._run_forever = run_forever
@@ -47,13 +47,13 @@ class NvrAlarmMonitoringTask:
 
     @retry(retry=retry_if_exception_type(Exception), wait=wait_fixed(RETRY_WAIT))
     async def run(self) -> None:
-        """Monitor and process alarm/alert chunks from NVR Alarm Stream."""
+        """Monitor and process alarm/alert chunks from NVR Alarm StreamType."""
         self._log.info('Starting NVR Alarm Monitoring Task')
         try:
             await self._process_chunks()
         except ChunkLoopError:
             self._log.error(
-                'Unexpectedly exited from NVR "%s" Alert Stream chunk processing loop. '
+                'Unexpectedly exited from NVR "%s" Alert StreamType chunk processing loop. '
                 'Retrying in %s seconds...',
                 self._host,
                 self.RETRY_WAIT,
@@ -61,7 +61,7 @@ class NvrAlarmMonitoringTask:
             raise
         except ConnectError:
             self._log.error(
-                'Failed to connect to NVR "%s" Alert Stream. Retrying in %s seconds...',
+                'Failed to connect to NVR "%s" Alert StreamType. Retrying in %s seconds...',
                 self._host,
                 self.RETRY_WAIT,
             )
@@ -76,7 +76,7 @@ class NvrAlarmMonitoringTask:
             raise
 
     async def _process_chunks(self) -> None:
-        """Process chunks received from NVR Alert Stream."""
+        """Process chunks received from NVR Alert StreamType."""
         async for chunk in self._api.alert_stream():
             self._log.debug('Alert chunk from NVR "%s": %s', self._host, chunk)
             if not chunk:
@@ -101,7 +101,7 @@ class NvrAlarmMonitoringTask:
             raise RuntimeError(f'Could not find NVR channel name in chunk: "{chunk}"')
         return self._channel_name_to_cam_map[channel_name]
 
-    def _send_alerts(self, cam: HikvisionCam, detection_type: Detection) -> None:
+    def _send_alerts(self, cam: HikvisionCam, detection_type: DetectionType) -> None:
         self._log.info(
             '[%s - %s] Sending "%s" alerts', cam.id, cam.description, detection_type
         )
