@@ -1,24 +1,26 @@
-import abc
 import asyncio
 import logging
-from typing import TYPE_CHECKING
+from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING, Final
 
-from addict import Dict
-
+from hikcamerabot.config.schemas.main_config import BaseDVRStorageUploadConfSchema
 from hikcamerabot.enums import DvrUploadType
 
 if TYPE_CHECKING:
     from hikcamerabot.camera import HikvisionCam
     from hikcamerabot.services.stream.dvr.file_wrapper import DvrFile
 
-_UPLOAD_RETRY_WAIT = 5
+_UPLOAD_RETRY_WAIT: Final[int] = 5
 
 
-class AbstractDvrUploadTask(metaclass=abc.ABCMeta):
+class AbstractDvrUploadTask(ABC):
     UPLOAD_TYPE: DvrUploadType | None = None
 
     def __init__(
-        self, cam: 'HikvisionCam', conf: Dict, queue: asyncio.Queue['DvrFile']
+        self,
+        cam: 'HikvisionCam',
+        conf: BaseDVRStorageUploadConfSchema,
+        queue: asyncio.Queue['DvrFile'],
     ) -> None:
         self._log = logging.getLogger(self.__class__.__name__)
         self._cam = cam
@@ -26,6 +28,6 @@ class AbstractDvrUploadTask(metaclass=abc.ABCMeta):
         self._conf = conf
         self._queue = queue
 
-    @abc.abstractmethod
+    @abstractmethod
     async def run(self) -> None:
         pass
