@@ -1,6 +1,8 @@
 """Telegram Livestream module."""
 
-from hikcamerabot.enums import Stream, VideoEncoder
+from typing import Literal
+
+from hikcamerabot.enums import StreamType, VideoEncoderType
 from hikcamerabot.services.stream.abstract import (
     AbstractExternalLivestreamService,
 )
@@ -9,10 +11,10 @@ from hikcamerabot.services.stream.abstract import (
 class TelegramStreamService(AbstractExternalLivestreamService):
     """Telegram Livestream Service Class."""
 
-    name = Stream.TELEGRAM
+    NAME: Literal[StreamType.TELEGRAM] = StreamType.TELEGRAM
 
     def _generate_transcode_cmd(
-        self, cmd_tpl: str, cmd_transcode: str, enc_codec_name: VideoEncoder
+        self, cmd_tpl: str, cmd_transcode: str, enc_codec_name: VideoEncoderType
     ) -> None:
         try:
             inner_args = self._cmd_gen_dispatcher[enc_codec_name]()
