@@ -1,6 +1,8 @@
 """YouTube Livestream module."""
 
-from hikcamerabot.enums import Stream, VideoEncoder
+from typing import Literal
+
+from hikcamerabot.enums import StreamType, VideoEncoderType
 from hikcamerabot.services.stream.abstract import (
     AbstractExternalLivestreamService,
 )
@@ -9,15 +11,16 @@ from hikcamerabot.services.stream.abstract import (
 class YouTubeStreamService(AbstractExternalLivestreamService):
     """YouTube Livestream Service Class."""
 
-    name = Stream.YOUTUBE
+    NAME: Literal[StreamType.YOUTUBE] = StreamType.YOUTUBE
 
     def _generate_transcode_cmd(
-        self, cmd_tpl: str, cmd_transcode: str, enc_codec_name: VideoEncoder
+        self, cmd_tpl: str, cmd_transcode: str, enc_codec_name: VideoEncoderType
     ) -> None:
         try:
             inner_args = self._cmd_gen_dispatcher[enc_codec_name]()
         except KeyError:
             inner_args = ''
+
         inner_args = cmd_transcode.format(inner_args=inner_args)
         self._cmd = cmd_tpl.format(
             output=self._generate_output(), inner_args=inner_args
