@@ -1,1 +1,1 @@
-__version__ = '1.8'
+__version__ = '1.9'
